Move stack pool to stack.rs
