+++ src/execution_engine.rs
@@ -13,46 +13,11 @@ use crate::{error::OrReturn, function::Function};
 use std::cell::UnsafeCell;
 use std::rc::Rc;
 
+use self::stack::StackPool;
+
+pub mod stack;
+
 pub type Stack<'a, T> = &'a mut [T];
-
-pub struct StackPool<T: Default> {
-    stack: Vec<T>,
-    base: usize,
-}
-
-impl<T: Default> StackPool<T> {
-    pub fn with_capacity(capacity: usize) -> StackPool<T> {
-        let mut stack = Vec::with_capacity(capacity);
-        for _ in 0..capacity {
-            stack.push(Default::default());
-        }
-        StackPool { stack, base: 0 }
-    }
-
-    pub fn request<'a>(&mut self, capacity: usize) -> &'a mut [T] {
-        if self.base + capacity >= self.stack.len() {
-            panic!("Stack overflow {} / {}", self.base, self.stack.len());
-        }
-
-        unsafe {
-            let ptr = self.stack.as_mut_ptr().add(self.base);
-
-            self.base += capacity;
-            let slice = std::slice::from_raw_parts_mut(ptr, capacity);
-            slice
-        }
-    }
-
-    pub fn release(&mut self, capacity: usize) {
-        self.base -= capacity;
-    }
-}
-
-impl<T: Default> Default for StackPool<T> {
-    fn default() -> Self {
-        Self::with_capacity(10000)
-    }
-}
 
 pub struct ExecutionEngine<TS: TypeSystem> {
     pub(crate) num_globals: usize,
@@ -136,7 +101,7 @@ impl<TS: TypeSystem> ExecutionEngine<TS> {
         mut args: impl FnMut(&mut ExecutionEngine<TS>) -> Result<TS::Value, FreightError>,
         arg_count: usize,
     ) -> Result<TS::Value, FreightError> {
-        let stack = self.stack.request(func.stack_size);
+        let mut stack = self.stack.request(func.stack_size);
         if !func.arg_count.valid_arg_count(arg_count) {
             return Err(FreightError::IncorrectArgumentCount {
                 expected_min: func.arg_count.min(),
@@ -180,8 +145,8 @@ impl<TS: TypeSystem> ExecutionEngine<TS> {
         }
         let function = self.get_function(func.location);
         let value = match &func.function_type {
-            FunctionType::CapturingRef(captures) => function.call(self, stack, captures),
-            FunctionType::Static => function.call(self, stack, &[]),
+            FunctionType::CapturingRef(captures) => function.call(self, &mut stack, captures),
+            FunctionType::Static => function.call(self, &mut stack, &[]),
             FunctionType::CapturingDef(_) => Err(FreightError::InvalidInvocationTarget),
             FunctionType::Native(_) => unreachable!("Native function already handled"),
         };
